requirements.txt
Flask
requests
python-dotenv
google-generativeai
matplotlib
pandas
